refactor: Coin.java를 domain 패키지로 이동
